feat(client/pages/Profile/AdminDashboard): edit All Sneakers section
