+++ client/src/components/index.js
@@ -22,4 +22,4 @@ export * from './PriceFilter/PriceFilter'
 export * from './SearchBar/SearchBar'
 export * from './CurrentFilters/CurrentFilters'
 export * from './UserModal/UserModal'
-export * from './AdminCards/AdminCards'
+export * from './AdminCard/AdminCard'
