+++ client/src/pages/Profile/AdminDashboard/AllSneakers/AllSneakers.jsx
@@ -1,22 +1,30 @@
 import { useGetShoesQuery } from '../../../../redux/services/filteredShoes'
-import { AdminCard } from '../../../../components'
+import { AdminCard, Loader, NotFound } from '../../../../components'
 import { useSelector } from 'react-redux'
+import styles from './AllSneakers.module.css'
 
-export default function AllSneakers() {
-  let slicedData = []
-  const filterState = useSelector(state => state.filter)
-  const { data } = useGetShoesQuery(filterState)
-  if (data) {
-    // const filteredData = data.filter((shoe) => shoe.STATUS === true)
-    // if ((page - 1) * pageSize > data.length) dispatch(setPage(1))
-    // dispatch(setTotalEntries(data.length))
-    slicedData = data
-  }
-  return (
-    <div>
-      {slicedData.map(shoe => (
-        <AdminCard key={shoe.id} shoe={shoe} />
-      ))}
-    </div>
-  )
+export function AllSneakers() {
+	const filterState = useSelector(state => state.filter)
+	const { data, isLoading, error } = useGetShoesQuery(filterState)
+
+	if (isLoading)
+		return (
+			<div className={styles.mt50vh}>
+				<Loader />
+			</div>
+		)
+
+	if (error)
+		return (
+			<div className={styles.mt50vh}>
+				<NotFound />
+			</div>
+		)
+	return (
+		<div className={styles.adminCardsContainer}>
+			{data?.map(shoe => (
+				<AdminCard key={shoe.id} shoe={shoe} />
+			))}
+		</div>
+	)
 }
