+++ client/src/pages/Profile/AdminDashboard/AllSneakers/AllSneakers.module.css
@@ -1,0 +1,6 @@
+.adminCardsContainer {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    column-gap: 1rem;
+}
