+++ client/src/pages/Profile/AdminDashboard/CreateSneaker/CreateSneaker.jsx
@@ -1,9 +1,3 @@
-import React from 'react'
-
-function CreateSneaker() {
-  return (
-    <div>CreateSneaker</div>
-  )
+export function CreateSneaker() {
+	return <div>CreateSneaker</div>
 }
-
-export default CreateSneaker
