+++ client/src/pages/Profile/AdminDashboard/EditSneaker/EditSneaker.module.css
@@ -1,0 +1,161 @@
+.mainContainer {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+}
+
+.loader,
+.error {
+	margin-top: 50vh;
+	margin-bottom: 100px;
+}
+
+.shoeContainer {
+	font-family: Poppins;
+	color: #242424;
+	display: flex;
+	flex-direction: row;
+}
+
+.allImagesContainer {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+}
+
+.separateImagesContainer {
+	display: flex;
+	flex-wrap: wrap;
+}
+
+.imageElement {
+	width: 125px;
+}
+
+.carouselContainer {
+	display: flex;
+	justify-content: center;
+	width: 40rem;
+}
+
+.descriptionContainer {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	flex-wrap: wrap;
+	padding: 0 3%;
+}
+
+.title {
+	font-size: 20px;
+}
+
+.code {
+	font-size: 12px;
+}
+
+.shoePrice {
+	font-family: Poppins, sans-serif;
+	font-size: 30px;
+	font-weight: 700;
+	margin: 20px 0 10px;
+}
+
+.sizesContainer {
+	margin: 10px 0;
+	padding: 10px 0 15px;
+	border-top: 1px solid #00000066;
+	border-bottom: 1px solid #00000066;
+}
+
+.sizesTitle {
+	font-size: 14px;
+	font-weight: 800;
+}
+
+.sizesChecks {
+	display: flex;
+	align-items: center;
+
+	flex-wrap: wrap;
+	gap: 16px;
+	padding: 10px 15px 5px;
+}
+
+.sizeCheck,
+.sizeChecked,
+.sizeDisabled {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+
+	width: 32px;
+	border: 2px solid #101113;
+	border-radius: 50%;
+	padding: 4px 4px 3px 4px;
+
+	text-align: center;
+	font-size: 0.85rem;
+	font-weight: 500;
+	color: #101113;
+
+	user-select: none;
+	cursor: pointer;
+	transition: 0.2s;
+}
+
+.sizeCheck {
+	background-color: white;
+}
+
+.sizeChecked {
+	font-weight: 900;
+	background-color: #101113;
+	color: #fff;
+
+	transition: 0.2s;
+}
+
+.sizeCheck:hover {
+	border: 2px solid #101113;
+	color: #101113;
+	/* box-shadow: inset 0 0 2px 1px #101113; */
+	transition: 0.2s;
+}
+
+.sizeDisabled {
+	opacity: 0.2;
+	cursor: default;
+}
+
+.addToCartButton {
+	display: block;
+	background-color: #ccc;
+	font-family: Poppins;
+	font-size: 22px;
+	font-weight: 600;
+
+	width: 80%;
+	max-width: 350px;
+	height: 55px;
+	margin: 10px auto 0;
+	padding: 4px 8px 0;
+	border-radius: 3px;
+}
+
+.addCartAlert {
+	display: block;
+	width: 350px;
+	margin: 0 auto;
+}
+
+.description {
+	margin: 40px 0;
+}
+
+.descriptionText {
+	margin-top: 20px;
+	line-height: 1.5;
+}
