+++ client/src/pages/Profile/AdminDashboard/Home/Home.jsx
@@ -1,9 +1,3 @@
-import React from 'react'
-
-function Home() {
-  return (
-    <div>Home</div>
-  )
+export function Home() {
+	return <div>Home</div>
 }
-
-export default Home
