+++ client/src/pages/Profile/AdminDashboard/Orders/Orders.jsx
@@ -1,9 +1,3 @@
-import React from 'react'
-
-function Orders() {
-  return (
-    <div>Orders</div>
-  )
+export function Orders() {
+	return <div>Orders</div>
 }
-
-export default Orders
