+++ client/src/pages/Profile/AdminDashboard/index.js
@@ -1,0 +1,5 @@
+export * from './EditSneaker/EditSneaker'
+export * from './AllSneakers/AllSneakers'
+export * from './CreateSneaker/CreateSneaker'
+export * from './Home/Home'
+export * from './Orders/Orders'
