+++ client/src/pages/index.js
@@ -10,4 +10,3 @@ export * from './Profile/Profile'
 export * from './Page404/Page404'
 export * from './Checkout/Checkout'
 export * from './Payment/PaymentSuccess'
-export * from './AdminDetail/AdminDetail'
